rename: svgIcon, shadcn ui 폴더 구조 변경 적용
