+++ src/components/common/molecule/Dropdown.tsx
@@ -5,7 +5,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from '@components/ui/select';
+} from '@components/common/atom/select';
 
 const Dropdown = () => {
   return (
